Sorting Algorithm
This Bubble Sort program with Pass. The first pass will push the largest element in the array to the last position in the index. The second Pass will push the second largest element to second to last position in the index. Until all the values are sorted in ascending order. There will be as many passes as there are elements the user inputs.